requirements.txt
python-dotenv
fastapi
httpx
loguru
leptonai
openai
duckduckgo-search
langchain-community
langchain-chroma
langchain-huggingface